requirements.txt
python-dotenv>=1.0
twelvelabs>=0.4,<0.5
requests>=2.31.0
opencv-python-headless>=4.9
Flask>=2.3.0
flask-cors>=4.0.0
google-generativeai>=0.3.0 